Estimate Edit updated

+++ app/src/main/java/com/dss/dsboxplus/data/repo/response/EstimateListResponse.kt
@@ -82,7 +82,7 @@ data class DataItem(
     val lengthCmField: @RawValue Any? = null,
 
     @field:SerializedName("businessid")
-    val businessid: Double? = null,
+    val businessid: Long? = null,
 
     @field:SerializedName("toprate")
     val toprate: @RawValue Any? = null,
